Refactor the specs
