+++ spec/features/user/account/avatar/remove_avatar_spec.rb
