+++ spec/features/user/my_posts/create_my_post_spec.rb
@@ -1,19 +1,19 @@
 require "rails_helper"
 
-feature "User creates a Post" do
+feature "Create a Post" do
   include_context "current user signed in"
   let(:post_attributes) { attributes_for(:post) }
 
   before { visit new_my_post_path }
 
-  scenario "with title" do
+  scenario "User creates a post with a title" do
     fill_form_and_submit(:post, post_attributes)
 
     expect(current_path).to eq post_path(Post.last)
     expect(page).to have_content "Post was successfully created."
   end
 
-  scenario "with no title" do
+  scenario "User created a post without a title" do
     fill_form_and_submit(:post, post_attributes.except(:title))
 
     expect(current_path).to eq my_posts_path
