+++ spec/features/user/my_posts/destroy_my_post_spec.rb
@@ -1,11 +1,11 @@
 require "rails_helper"
 
-feature "Delete a post" do
+feature "Delete the post" do
   include_context "current user signed in"
   include_context "current user's post"
   let(:another_user) { create(:user) }
 
-  scenario "Owner deletes the post" do
+  scenario "Owner removes the post" do
     visit post_path(post)
     click_on "Delete the Post"
 
@@ -13,7 +13,7 @@ feature "Delete a post" do
     expect(page).to have_content "Post was successfully destroyed."
   end
 
-  scenario "Another User deletes the post" do
+  scenario "Another User removes the post" do
     login_as another_user
     visit post_path(post)
