+++ spec/features/user/posts/see_a_post_spec.rb
