+++ spec/features/user/posts/see_all_posts_spec.rb
