+++ spec/features/user/user_statistics_spec.rb
