+++ spec/features/visitor/posts/create_a_post_spec.rb
