+++ spec/features/visitor/posts/see_posts_spec.rb
