+++ spec/features/visitor/posts/show_a_post_spec.rb
